Parti 2

+++ srcs/ft_strmapi.c
@@ -6,14 +6,14 @@
 /*   By: dmoureu- <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/24 20:30:27 by dmoureu-          #+#    #+#             */
-/*   Updated: 2015/11/24 20:34:07 by dmoureu-         ###   ########.fr       */
+/*   Updated: 2015/11/25 16:44:02 by dmoureu-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdlib.h>
 #include <libft.h>
 
-char *ft_strmapi(char const *s, char (*f)(unsigned int, char))
+char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
 	unsigned int	i;
 	int				len;
@@ -24,7 +24,7 @@ char *ft_strmapi(char const *s, char (*f)(unsigned int, char))
 	str = (char *)malloc(sizeof(char) * (len + 1));
 	if (!str)
 		return (NULL);
-	while (i < len)
+	while ((int)i < len)
 	{
 		str[i] = f(i, s[i]);
 		i++;
